.virtualenv.requirements.txt
#
# Requirement files for Python Virtual Environments
#

# Shell interface for docopt, the command-line interface description language
docopt

# YAML file parsing
PyYAML

# Daemon
apscheduler

# kiwi
kiwi

# Access to kafka
kafka-python

# Transport schema
cerberus

# Network access
requests

# Process Management
psutil

# SSH
paramiko
